requirements.txt
google-cloud-storage==7.19.0
google-generativeai==0.7.2
python-dotenv==1.0.0
